requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
faiss-cpu==1.8.0
numpy==1.26.4
python-multipart==0.0.9
google-generativeai==0.5.4
pypdf==4.2.0
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.7.1
